Added Redis server for config
Rather than the previous setup where dynamic data
(sync times, eq coefficients, etc) is saved in a file
in the config directory, this data is now stored in a redis
backend.

Redis also obsoletes the need to pass each python script
a config file.

A dedicated python script scripts/generic/write_config_redis.py
can be used to push a configuration (from a config file) to
redis. Other scripts will then use this configuration when they run.
This allows distributed processes, potentially on different servers,
to avoid clashing configurations.

The configuration file is now YAML, which has the benefit of
automatic type parsing, and neat, nested structures.

The contents of the config file should mostly be self-explanatory,
and in any case the parameters are largely thee same as the old format
(with the addition of a redis gateway address). By default, since scripts
are no longer run with config files, the redis server is assumed to have
the host name ami_redis_host, which should be set in the hosts file
or otherwise provided over DNS