Modify X-engine class
It now has the same config file -> attribute generation scheme
as the F-Engine. I.e., in the config file, keys in the Xengine
section end up as XEngine instance attirbutes, and can be overridden
on a per-engine basis.

Add loads of methods for xenging ctrl-sw manipulation and other
packetized correlator methods