Add support for multicast
The config file now allows a "multicast_order" option.

If set to 0, multicast is not used.
If set to 1, each roach 10gbe link receives on a single multicast address. Other
users can potentially subscribe to the address of one of the roach ports to get a
duplicate of the data stream.
If set to 2, each roach link receives on 2 consecutive addresses. One of these can then
be used to mirror subsets of data to another user.

Currently, multicast subscription is not supported by the firmware (cpu tge interface has
not been compiled in), so the switch in the correlator needs to be configured with
static multicast routes.